requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
pymysql
sqlalchemy
streamlit-autorefresh